tests: Move `GET /crates/:name` tests to `krate::show` module
